VIITE-3037 Import changes due to renaming old fi.liikennevirasto.digiroad2.dao to fi.vaylavirasto.viite.dao
