+++ viite-backend/api/src/test/scala/fi/liikennevirasto/digiroad2/ViiteApiSpec.scala
@@ -4,10 +4,10 @@ import fi.liikennevirasto.digiroad2.client.kgv.{KgvRoadLink, KgvRoadLinkClient}
 import fi.liikennevirasto.digiroad2.service.RoadLinkService
 import fi.liikennevirasto.digiroad2.user.{Configuration, User}
 import fi.liikennevirasto.digiroad2.Digiroad2Context._
-import fi.liikennevirasto.digiroad2.dao.ComplementaryLinkDAO
 import fi.liikennevirasto.viite.{NodesAndJunctionsService, PreFillInfo, ProjectService, RoadAddressService, RoadNameService, RoadNameSource, RoadNetworkValidator, ViiteVkmClient}
 import fi.liikennevirasto.viite.dao.{ProjectLinkDAO}
 import fi.liikennevirasto.viite.util.{DigiroadSerializers, JsonSerializer, runWithRollback}
+import fi.vaylavirasto.viite.dao.ComplementaryLinkDAO
 import fi.vaylavirasto.viite.geometry.{BoundingRectangle, Point}
 import fi.vaylavirasto.viite.model.{AdministrativeClass, LifecycleStatus, LinkGeomSource, RoadLink, TrafficDirection}
 import org.joda.time.format.{DateTimeFormat, DateTimeFormatter}
