+++ viite-backend/conf/env.properties
@@ -13,8 +13,8 @@ conversion.bonecp.username=drkonv
 env=Local
 eventBus=fi.liikennevirasto.digiroad2.DigiroadEventBus
 importOnlyCurrent=false
-municipalityProvider=fi.liikennevirasto.digiroad2.dao.PostGISMunicipalityProvider
-userProvider=fi.liikennevirasto.digiroad2.dao.PostGISUserProvider
+municipalityProvider=fi.vaylavirasto.viite.dao.PostGISMunicipalityProvider
+userProvider=fi.vaylavirasto.viite.dao.PostGISUserProvider
 vkmUrl=https://api.vaylapilvi.fi
 #vkmApiKey can be found in Viite confluence password storage page
 vkmApiKey=
