+++ viite-backend/database/src/main/scala/fi/vaylavirasto/viite/dao/PostGISMunicipalityProvider.scala
@@ -1,4 +1,4 @@
-package fi.liikennevirasto.digiroad2.dao
+package fi.vaylavirasto.viite.dao
 
 import fi.liikennevirasto.digiroad2.postgis.PostGISDatabase
 import slick.driver.JdbcDriver.backend.Database
