+++ viite-backend/database/src/main/scala/fi/vaylavirasto/viite/dao/SequenceResetterDAO.scala
@@ -1,4 +1,4 @@
-package fi.liikennevirasto.digiroad2.dao
+package fi.vaylavirasto.viite.dao
 
 import slick.driver.JdbcDriver.backend.Database.dynamicSession
 import slick.jdbc.StaticQuery
