+++ viite-backend/database/src/test/scala/fi/vaylavirasto/viite/dao/PostGISMunicipalityProviderSpec.scala
@@ -1,4 +1,4 @@
-package fi.liikennevirasto.digiroad2.dao
+package fi.vaylavirasto.viite.dao
 
 import org.scalatest.{FunSuite, Matchers, Tag}
