+++ viite-backend/viite-main/src/main/scala/fi/liikennevirasto/viite/dao/NodeDAO.scala
@@ -2,10 +2,10 @@ package fi.liikennevirasto.viite.dao
 
 import java.sql.Timestamp
 import com.github.tototoshi.slick.MySQLJodaSupport._
-import fi.liikennevirasto.digiroad2.dao.Sequences
 import fi.liikennevirasto.digiroad2.postgis.PostGISDatabase
 import fi.liikennevirasto.digiroad2.util.LogUtils.time
 import fi.liikennevirasto.viite.NewIdValue
+import fi.vaylavirasto.viite.dao.Sequences
 import fi.vaylavirasto.viite.geometry.{BoundingRectangle, Point}
 import fi.vaylavirasto.viite.model.{NodePointType, NodeType}
 import org.joda.time.DateTime
