+++ viite-backend/viite-main/src/main/scala/fi/liikennevirasto/viite/dao/RoadNameDAO.scala
@@ -3,8 +3,8 @@ package fi.liikennevirasto.viite.dao
 import java.sql.{Date, Timestamp}
 
 import com.github.tototoshi.slick.MySQLJodaSupport._
-import fi.liikennevirasto.digiroad2.dao.Sequences
 import fi.liikennevirasto.digiroad2.user.User
+import fi.vaylavirasto.viite.dao.Sequences
 import org.joda.time.DateTime
 import org.joda.time.format.{DateTimeFormat, DateTimeFormatter}
 import org.slf4j.LoggerFactory
