+++ viite-backend/viite-main/src/main/scala/fi/liikennevirasto/viite/process/RoadwayFiller.scala
@@ -1,9 +1,9 @@
 package fi.liikennevirasto.viite.process
 
-import fi.liikennevirasto.digiroad2.dao.Sequences
 import fi.liikennevirasto.viite.NewIdValue
 import fi.liikennevirasto.viite.dao._
 import fi.liikennevirasto.viite.dao.TerminationCode.{NoTermination, Subsequent}
+import fi.vaylavirasto.viite.dao.Sequences
 import fi.vaylavirasto.viite.model.{AdministrativeClass, Discontinuity, RoadAddressChangeType, Track}
 import org.joda.time.DateTime
 import org.slf4j.LoggerFactory
