+++ viite-backend/viite-main/src/test/scala/fi/liikennevirasto/viite/dao/JunctionPointDAOSpec.scala
@@ -1,8 +1,8 @@
 package fi.liikennevirasto.viite.dao
 
-import fi.liikennevirasto.digiroad2.dao.Sequences
 import fi.liikennevirasto.digiroad2.postgis.PostGISDatabase
 import fi.liikennevirasto.viite.NewIdValue
+import fi.vaylavirasto.viite.dao.Sequences
 import fi.vaylavirasto.viite.model.{AdministrativeClass, BeforeAfter, Discontinuity, Track}
 import org.joda.time.DateTime
 import org.scalatest.{FunSuite, Matchers}
