+++ viite-backend/viite-main/src/test/scala/fi/liikennevirasto/viite/dao/RoadNameDAOSpec.scala
@@ -1,7 +1,7 @@
 package fi.liikennevirasto.viite.dao
 
-import fi.liikennevirasto.digiroad2.dao.Sequences
 import fi.liikennevirasto.digiroad2.postgis.PostGISDatabase
+import fi.vaylavirasto.viite.dao.Sequences
 import fi.vaylavirasto.viite.model.{AdministrativeClass, Discontinuity, Track}
 import org.scalatest.{FunSuite, Matchers}
 import org.joda.time.DateTime
